perf: derive Logs filter via useMemo + lazy-load recharts in MemberProfile
Two remaining real doctor findings worth fixing.

imports/ui/logs/Logs.tsx — `filter` was useState seeded by useEffect
that recomputed `buildFilter(actionInput, dateRange)` on every input
change. Pure derived state — convert to useMemo, drop the useState +
useEffect pair.

imports/ui/members/MemberProfile.tsx — recharts ships ~100 KB into
the initial bundle even though the AttendancePieChart only renders
inside an expanded member row that most users never open. Extract the
chart to its own file, import via React.lazy, and wrap each render
site in Suspense. The expanded row already has its own loading state
so the Suspense fallback only shows on the brief lazy-chunk fetch.

Skipped from the remaining ~65 doctor findings

React 18 auto-batches multiple setState calls within a single
microtask (.then callbacks, event handlers), so the "cascading-
set-state" and "prefer-useReducer" rules (9 + 7 findings) are
stylistic rather than functional in this codebase — refactoring them
into useReducer/object-state would add abstraction beyond what the
task requires, against the project's own coding guidelines.

The remaining findings split as:
  - ~30 stylistic (cascading-set-state, prefer-useReducer, generic
    handler names, inline styles, barrel imports)
  - ~10 false positives (Header.tsx document sync, hydration-on-non-SSR,
    state-in-early-returns, advanced-event-handler-refs where the
    handlers ARE memoised)
  - ~10 idiomatic React patterns flagged anyway (Section/Palette
    reset-on-change effects)

The full per-finding breakdown lives in the PR description.

Test plan

- [x] npm test — 311 passing, unchanged
- [x] npm run typecheck — 8 pre-existing palette errors, no new ones
- [ ] npm run e2e — will run before merge
